refactor selectors structure (#1022)
