+++ src/components/AccountAssets/AccountAssets.tsx
@@ -1,8 +1,10 @@
 import { CAIP19 } from '@shapeshiftoss/caip'
 import { Text } from 'components/Text'
-import { selectAssetByCAIP19 } from 'state/slices/assetsSlice/assetsSlice'
 import { AccountSpecifier } from 'state/slices/portfolioSlice/portfolioSlice'
-import { selectPortfolioAssetIdsByAccountIdExcludeFeeAsset } from 'state/slices/portfolioSlice/selectors'
+import {
+  selectAssetByCAIP19,
+  selectPortfolioAssetIdsByAccountIdExcludeFeeAsset
+} from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 
 import { Card } from '../Card/Card'
