+++ src/components/AssetAccounts/AssetAccounts.tsx
@@ -4,7 +4,7 @@ import { useTranslate } from 'react-polyglot'
 import { Card } from 'components/Card/Card'
 import { Text } from 'components/Text'
 import { AccountSpecifier } from 'state/slices/portfolioSlice/portfolioSlice'
-import { selectAccountIdsByAssetId } from 'state/slices/portfolioSlice/selectors'
+import { selectAccountIdsByAssetId } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 
 import { AssetAccountRow } from './AssetAccountRow'
