+++ src/components/SelectAssets/SelectAccount.tsx
@@ -6,9 +6,8 @@ import { useTranslate } from 'react-polyglot'
 import { useHistory, useLocation } from 'react-router'
 import { AssetAccountRow } from 'components/AssetAccounts/AssetAccountRow'
 import { SlideTransition } from 'components/SlideTransition'
-import { selectAssetByCAIP19 } from 'state/slices/assetsSlice/assetsSlice'
 import { AccountSpecifier } from 'state/slices/portfolioSlice/portfolioSlice'
-import { selectAccountIdsByAssetId } from 'state/slices/portfolioSlice/selectors'
+import { selectAccountIdsByAssetId, selectAssetByCAIP19 } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 
 import { SelectAssetRoutes } from './SelectAssetRouter'
