+++ src/components/StakingVaults/EarnOpportunities.tsx
@@ -7,8 +7,8 @@ import { NavLink } from 'react-router-dom'
 import { Card } from 'components/Card/Card'
 import { Text } from 'components/Text'
 import { useYearnVaults } from 'hooks/useYearnVaults/useYearnVaults'
-import { selectAssetByCAIP19 } from 'state/slices/assetsSlice/assetsSlice'
 import { AccountSpecifier } from 'state/slices/portfolioSlice/portfolioSlice'
+import { selectAssetByCAIP19 } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 
 import { EarnOpportunityRow } from './EarnOpportunityRow'
