+++ src/components/Trade/hooks/useTradeRoutes/useTradeRoutes.ts
@@ -3,7 +3,7 @@ import { useCallback, useEffect } from 'react'
 import { useFormContext } from 'react-hook-form'
 import { useSelector } from 'react-redux'
 import { useHistory } from 'react-router-dom'
-import { selectAssets } from 'state/slices/assetsSlice/assetsSlice'
+import { selectAssets } from 'state/slices/selectors'
 
 import { TradeState } from '../../Trade'
 import { TradeActions, useSwapper } from '../useSwapper/useSwapper'
