+++ src/components/TxHistory.tsx
@@ -9,10 +9,12 @@ import { TransactionRow } from 'components/Transactions/TransactionRow'
 import { useWallet } from 'context/WalletProvider/WalletProvider'
 import { useInfiniteScroll } from 'hooks/useInfiniteScroll/useInfiniteScroll'
 import { useWalletSupportsChain } from 'hooks/useWalletSupportsChain/useWalletSupportsChain'
-import { selectAssetByCAIP19 } from 'state/slices/assetsSlice/assetsSlice'
 import { AccountSpecifier } from 'state/slices/portfolioSlice/portfolioSlice'
-import { selectAccountIdsByAssetId } from 'state/slices/portfolioSlice/selectors'
-import { selectTxIdsByFilter } from 'state/slices/txHistorySlice/txHistorySlice'
+import {
+  selectAccountIdsByAssetId,
+  selectAssetByCAIP19,
+  selectTxIdsByFilter
+} from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 
 type TxHistoryProps = {
