+++ src/context/TransactionsProvider/TransactionsProvider.tsx
@@ -4,9 +4,8 @@ import { useDispatch, useSelector } from 'react-redux'
 import { useChainAdapters } from 'context/ChainAdaptersProvider/ChainAdaptersProvider'
 import { useWallet } from 'context/WalletProvider/WalletProvider'
 import { walletSupportChain } from 'hooks/useWalletSupportsChain/useWalletSupportsChain'
-import { selectAssets } from 'state/slices/assetsSlice/assetsSlice'
 import { supportedAccountTypes } from 'state/slices/portfolioSlice/portfolioSlice'
-import { selectAccountIdByAddress } from 'state/slices/portfolioSlice/selectors'
+import { selectAccountIdByAddress, selectAssets } from 'state/slices/selectors'
 import { txHistory } from 'state/slices/txHistorySlice/txHistorySlice'
 import { store } from 'state/store'
