+++ src/pages/Accounts/Accounts.tsx
@@ -2,7 +2,7 @@ import { Flex, Heading, Stack } from '@chakra-ui/react'
 import { useSelector } from 'react-redux'
 import { Page } from 'components/Layout/Page'
 import { Text } from 'components/Text'
-import { selectPortfolioAccountIdsSortedFiat } from 'state/slices/portfolioSlice/selectors'
+import { selectPortfolioAccountIdsSortedFiat } from 'state/slices/selectors'
 
 import { AccountRowWithTokens } from './AccountRowWithTokens'
